+++ app/src/main/java/com/clipboarder/clipboarder/service/ClipboarderKeyboardService.kt
@@ -4,19 +4,9 @@ import android.content.Intent
 import android.inputmethodservice.InputMethodService
 import android.view.View
 import androidx.annotation.CallSuper
-import androidx.compose.foundation.ExperimentalFoundationApi
-import androidx.compose.foundation.background
-import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.height
-import androidx.compose.foundation.pager.HorizontalPager
-import androidx.compose.foundation.pager.rememberPagerState
 import androidx.compose.runtime.Composable
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.platform.AbstractComposeView
 import androidx.compose.ui.platform.ViewCompositionStrategy
-import androidx.compose.ui.unit.dp
 import androidx.lifecycle.Lifecycle
 import androidx.lifecycle.LifecycleOwner
 import androidx.lifecycle.ServiceLifecycleDispatcher
@@ -30,8 +20,7 @@ import androidx.savedstate.SavedStateRegistryOwner
 import androidx.savedstate.setViewTreeSavedStateRegistryOwner
 import com.clipboarder.clipboarder.data.repository.ContentRepository
 import com.clipboarder.clipboarder.data.repository.UserRepository
-import com.clipboarder.clipboarder.ui.screens.ime.clipboarder_keyboard.ClipboarderKeyboard
-import com.clipboarder.clipboarder.ui.screens.ime.language_keyboard.LanguageKeyboard
+import com.clipboarder.clipboarder.ui.screens.ime.ClipboarderIME
 import dagger.hilt.android.AndroidEntryPoint
 import javax.inject.Inject
 
@@ -109,35 +98,5 @@ class ClipboarderKeyboardService : InputMethodService(), LifecycleOwner,
     override fun onDestroy() {
         lifecycleDispatcher.onServicePreSuperOnDestroy()
         super.onDestroy()
-    }
-}
-
-@OptIn(ExperimentalFoundationApi::class)
-@Composable
-fun ClipboarderIME(
-    inputMethodService: InputMethodService,
-    userRepository: UserRepository,
-    contentRepository: ContentRepository
-) {
-    val pagerState = rememberPagerState(
-        pageCount = { 1000 },
-        initialPage = 1000 / 2
-    )
-
-    Box(
-        modifier = Modifier
-            .fillMaxWidth()
-            .height(288.dp)
-            .background(Color.DarkGray)
-    ) {
-        HorizontalPager(
-            state = pagerState,
-            modifier = Modifier.fillMaxWidth()
-        ) { page ->
-            when (page % 2) {
-                0 -> LanguageKeyboard(inputMethodService)
-                1 -> ClipboarderKeyboard(inputMethodService, userRepository, contentRepository)
-            }
-        }
     }
 }
