+++ app/src/main/java/com/clipboarder/clipboarder/ui/screens/ime/ClipboarderIMEViewModel.kt
@@ -1,0 +1,9 @@
+package com.clipboarder.clipboarder.ui.screens.ime
+
+import androidx.lifecycle.ViewModel
+import dagger.hilt.android.lifecycle.HiltViewModel
+import javax.inject.Inject
+
+@HiltViewModel
+class ClipboarderIMEViewModel @Inject constructor() : ViewModel() {
+}
